Remove cruft from landlab.grid.structured_quad (#1149)
* remove unused modules from landlab.grid.structured_quad

* remove unused node_in_link_ids, node_out_link_ids

* move structured_quad.links functions into overland_flow

* update docs for remove structured_quad subpackage

* use private versions of structured_quad functions

* add tests for graph freeze/thaw

* add nodes_at_link to StructuredQuadGraphTopology
