+++ landlab/graph/structured_quad/structured_quad.py
@@ -460,6 +460,12 @@ class StructuredQuadGraphTopology:
 
     @property
     @lru_cache()
+    @read_only_array
+    def nodes_at_link(self):
+        return self._layout.nodes_at_link(self.shape)
+
+    @property
+    @lru_cache()
     def horizontal_links(self):
         return self._layout.horizontal_links(self.shape)
 
@@ -471,7 +477,9 @@ class StructuredQuadGraphTopology:
     @property
     def corner_nodes(self):
         n_rows, n_cols = self.shape
-        return (n_rows * n_cols - 1, (n_rows - 1) * n_cols, 0, n_cols - 1)
+        return np.asarray(
+            (n_rows * n_cols - 1, (n_rows - 1) * n_cols, 0, n_cols - 1), dtype=int
+        )
 
     @property
     @lru_cache()
@@ -511,6 +519,10 @@ class StructuredQuadGraphExtras(StructuredQuadGraphTopology, Graph):
             patches=StructuredQuadLayoutCython.links_at_patch(self.shape),
             sort=sort,
         )
+
+    @property
+    def nodes_at_link(self):
+        return self.ds["nodes_at_link"].values
 
 
 class StructuredQuadGraph(StructuredQuadGraphExtras):
