+++ tests/components/overland_flow/test_dealmeida_overland_flow.py
@@ -8,10 +8,14 @@ import numpy as np
 
 from landlab import RasterModelGrid
 from landlab.components.overland_flow import OverlandFlow
-from landlab.grid.structured_quad.links import left_edge_horizontal_ids
+from landlab.components.overland_flow._links import horizontal_link_ids
 
 (_SHAPE, _SPACING, _ORIGIN) = ((32, 240), (25, 25), (0.0, 0.0))
 _ARGS = (_SHAPE, _SPACING, _ORIGIN)
+
+
+def _left_edge_horizontal_ids(shape):
+    return horizontal_link_ids(shape)[:, 0]
 
 
 def test_deAlm_name(deAlm):
@@ -51,7 +55,7 @@ def test_deAlm_analytical():
     grid.add_zeros("surface_water__depth", at="node")
     grid.add_zeros("topographic__elevation", at="node")
     grid.set_closed_boundaries_at_grid_edges(True, True, True, True)
-    left_inactive_ids = left_edge_horizontal_ids(grid.shape)
+    left_inactive_ids = _left_edge_horizontal_ids(grid.shape)
     deAlm = OverlandFlow(grid, mannings_n=0.01, h_init=0.001)
     time = 0.0
 
@@ -84,7 +88,7 @@ def test_deAlm_analytical_imposed_dt_short():
     grid.add_zeros("surface_water__depth", at="node")
     grid.add_zeros("topographic__elevation", at="node")
     grid.set_closed_boundaries_at_grid_edges(True, True, True, True)
-    left_inactive_ids = left_edge_horizontal_ids(grid.shape)
+    left_inactive_ids = _left_edge_horizontal_ids(grid.shape)
     deAlm = OverlandFlow(grid, mannings_n=0.01, h_init=0.001)
     time = 0.0
 
@@ -117,7 +121,7 @@ def test_deAlm_analytical_imposed_dt_long():
     grid.add_zeros("surface_water__depth", at="node")
     grid.add_zeros("topographic__elevation", at="node")
     grid.set_closed_boundaries_at_grid_edges(True, True, True, True)
-    left_inactive_ids = left_edge_horizontal_ids(grid.shape)
+    left_inactive_ids = _left_edge_horizontal_ids(grid.shape)
     deAlm = OverlandFlow(grid, mannings_n=0.01, h_init=0.001)
     time = 0.0
